# Improve performance of decoding