--- a/tests/version-number.rs
+++ b/tests/version-number.rs
@@ -1,5 +1,3 @@
-#![allow(non_fmt_panic)]
-
 #[test]
 fn test_readme_deps() {
     version_sync::assert_markdown_deps_updated!("README.md");